feat: Move MoreInfoScreen to kmpuiviews and Update Dependencies
This commit moves the `MoreInfoScreen` to the `kmpuiviews` module for multiplatform use, removes the old `InfoSettings` from `UIViews`, and updates related dependencies.

- **`MoreInfoScreen`:**
    - Moved `MoreInfoScreen` from `UIViews` to `kmpuiviews`.
    - Removed the old `InfoSettings` from `UIViews`.
    - Updated `NavGraph` to use the new `MoreInfoScreen` from `kmpuiviews`.
-   - Updated `MoreInfoScreen` to use the new kmp locations
+   - Updated `MoreInfoScreen` to use the new kmp locations.
    - Added `BuildKonfig` to handle prerelease checks.
-    - Updated `MoreInfoScreen` to show update dialog
+    - Updated `MoreInfoScreen` to show update dialog.
+- **Other**
+   - Added `buildKonfig` dependency
+   - Updated `BuildKonfig` to handle prerelease checks.
    - General cleanup.
